+++ shima.js
@@ -1,5 +1,13 @@
 'use strict'
 process.title = 'Shimakaze-chan'
+
+try {
+  require('./config.json')
+} catch (e) {
+  console.log('Config file not found, make one using the example and restart WildBeast.')
+  process.exit()
+}
+
 var Discordie = require('discordie')
 var Event = Discordie.Events
 var bot
@@ -12,7 +20,6 @@ var datacontrol = runtime.datacontrol
 var argv = require('minimist')(process.argv.slice(2))
 var Config
 var restarted = false
-var called = false
 
 Logger.info('Initializing...')
 
@@ -39,26 +46,18 @@ if (argv.forceupgrade) {
       Logger.error(e)
     })
   }
-}
-
-if (!argv.forceupgrade) {
-  try {
-    require('fs').readFileSync('./runtime/initial.txt')
-    start()
-  } catch (e) {
-    if (!argv.noinitial) {
-      runtime.internal.init.initial().then(function () {
-        start()
-      })
-    } else {
-      Logger.debug('Skipped init via argv')
-      require('fs').writeFileSync('./runtime/initial.txt', 'Initial setup skipped.')
-      start()
-    }
-  }
+} else {
+  start()
 }
 
 bot.Dispatcher.on(Event.GATEWAY_READY, function () {
+  runtime.internal.versioncheck.versionCheck(function (err, res) {
+    if (err) {
+      Logger.error('Version check failed, ' + err)
+    } else if (res) {
+      Logger.info(`Version check: ${res}`)
+    }
+  })
   Logger.info('Ready to start!')
   Logger.info(`Logged in as ${bot.User.username}#${bot.User.discriminator} (ID: ${bot.User.id}) and serving ${bot.Users.length} users in ${bot.Guilds.length} servers.`)
 })
@@ -102,7 +101,18 @@ bot.Dispatcher.on(Event.MESSAGE_CREATE, function (c) {
         return // ignore JS build-in array functions
       }
       Logger.info(`Executing <${c.message.resolveContent()}> from ${c.message.author.username}`)
-      if (!c.message.isPrivate) {
+      if (commands[cmd].level === 'master') {
+        if (Config.permissions.master.indexOf(c.message.author.id) > -1) {
+          try {
+            commands[cmd].fn(c.message, suffix, bot)
+          } catch (e) {
+            c.message.channel.sendMessage('An error occured while trying to process this command, you should let the bot author know. \n```' + e + '```')
+            Logger.error(`Command error, thrown by ${commands[cmd].name}: ${e}`)
+          }
+        } else {
+          c.message.channel.sendMessage('This command is only for the bot owner.')
+        }
+      } else if (!c.message.isPrivate) {
         timeout.check(commands[cmd], c.message.guild.id, c.message.author.id).then((y) => {
           if (y !== true) {
             datacontrol.customize.reply(c.message, 'timeout').then((x) => {
@@ -149,8 +159,10 @@ bot.Dispatcher.on(Event.MESSAGE_CREATE, function (c) {
               } else {
                 datacontrol.customize.reply(c.message, 'permissions').then((u) => {
                   if (u === 'default') {
-                    var reason = (r > 4) ? '**This is a master user only command**, ask the bot owner to add you as a master user if you really think you should be able to use this command.' : 'Ask the server owner to modify your level with `setlevel`.'
-                    c.message.channel.sendMessage('You have no permission to run this command!\nYou need level ' + commands[cmd].level + ', you have level ' + r + '\n' + reason)
+                    if (r > -1 && !commands[cmd].hidden) {
+                      var reason = (r > 4) ? '**This is a master user only command**, ask the bot owner to add you as a master user if you really think you should be able to use this command.' : 'Ask the server owner to modify your level with `setlevel`.'
+                      c.message.channel.sendMessage('You have no permission to run this command!\nYou need level ' + commands[cmd].level + ', you have level ' + r + '\n' + reason)
+                    }
                   } else {
                     c.message.channel.sendMessage(u.replace(/%user/g, c.message.author.username).replace(/%server/g, c.message.guild.name).replace(/%channel/, c.message.channel.name).replace(/%nlevel/, commands[cmd].level).replace(/%ulevel/, r))
                   }
@@ -195,7 +207,11 @@ bot.Dispatcher.on(Event.GUILD_MEMBER_ADD, function (s) {
   datacontrol.customize.check(s.guild).then((r) => {
     if (r === 'on') {
       datacontrol.customize.reply(s, 'welcome').then((x) => {
-        s.guild.generalChannel.sendMessage(x.replace(/%user/g, s.member.username).replace(/%server/g, s.guild.name))
+        if (x === 'default') {
+          s.guild.generalChannel.sendMessage(`Welcome ${s.member.username} to ${s.guild.name}!`)
+        } else {
+          s.guild.generalChannel.sendMessage(x.replace(/%user/g, s.member.username).replace(/%server/g, s.guild.name))
+        }
       }).catch((e) => {
         Logger.error(e)
       })
@@ -247,13 +263,6 @@ bot.Dispatcher.on(Event.DISCONNECTED, function (e) {
 })
 
 function start () {
-  runtime.internal.versioncheck.versionCheck(function (err, res) {
-    if (err) {
-      Logger.error('Version check failed, ' + err)
-    } else if (res) {
-      Logger.info(res)
-    }
-  })
   try {
     Config = require('./config.json')
   } catch (e) {